+++ src/filters/MinFilterNTest.java
@@ -20,31 +20,31 @@ public class MinFilterNTest {
 	
 	@Test
 	public void testInputOutput() throws InvalidObjectException {
-		Double input = (double) 2;
+		Double input = 2d;
 		MinFilter filter = new MinFilterN(2);
 		
 		filter.addInput(input);
 		assertEquals("Only value should be output", input, filter.getOutput());
 		
-		input = (double) 3;
+		input = 3d;
 		filter.addInput(input);
 		assertNotEquals("New greater value shouldn't be output", input, filter.getOutput());
 		
-		input = (double) 4;
+		input = 4d;
 		filter.addInput(input);
 		assertNotEquals("Old value should still be output", input, filter.getOutput());
 		
-		input = (double) 1;
+		input = 1d;
 		filter.addInput(input);
 		assertEquals("Old max should be popped due to age", input, filter.getOutput());
 	}
 
 	@Test
 	public void testReset() throws InvalidObjectException {
-		Double reset = (double) 0;
+		Double reset = 0d;
 		MinFilterN filter = new MinFilterN(2);
 		
-		filter.addInput((double) 27);
+		filter.addInput(27d);
 		filter.reset(reset);
 		
 		assertEquals("Output should be the reset value", reset, filter.getOutput());
